requirements.txt
argparse
selenium
python-dotenv